docker/legacy-system-example/src/test/java/dev/codesoapbox/legacysystemexample/authentication/infrastructure/repositories/InMemoryUserRepositoryTest.java
package dev.codesoapbox.legacysystemexample.authentication.infrastructure.repositories;

import dev.codesoapbox.legacysystemexample.authentication.domain.model.UserData;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.util.List;
import java.util.Optional;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

class InMemoryUserRepositoryTest {

    private InMemoryUserRepository repository;

    @BeforeEach
    void setUp() {
        repository = new InMemoryUserRepository();
    }

    @Test
    void shouldFindByUsername() {
        Optional<UserData> user = repository.findByUsername("lucy");

        assertTrue(user.isPresent());
    }

    @Test
    void shouldFindAll() {
        List<UserData> users = repository.findAll();

        assertFalse(users.isEmpty());
    }
}